code refacored for client id api for posthog
